papers/samarajeewaArtificialIntelligenceFramework2024/LEAP-Baseline-Trigger/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-queue
pyodbc
dynaconf
pandas
pytz

